sample_requirements.txt
# Without redis, celery, google-generativeai, docling
# please install django-cors-headers (I forgot)

asgiref==3.8.1
Django==5.1.3
django-environ==0.11.2
django-soft-delete==1.0.15
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pdf2image==1.17.0
pillow==11.0.0
PyJWT==2.9.0
PyMuPDF==1.24.13
sqlparse==0.5.2
tzdata==2024.2

